Restructure panels
